Refactor web crawler to better manage memory
A lot in this:

* Used named functions instead of storing functions in variables.
* Use a global config object rather than individual global vars.
* Destory Elasticsearch object after every use.
* Limit node-crawler to 100 crawls before regenerating object.
* Use promises with ES rather than callbacks.
* Improve resove/reject handling and logging throughout.

This appears to improve the management of memory significantly.

There is still a large number of URLs that will be tracked and
stored, so the total memory footprint of the app will grow over
time. I'm hoping this does enough to give us a longer run time
than several hours though. :)